requirements.txt
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
flask~=3.0.3
python-dotenv~=1.0.0
werkzeug~=3.0.2